requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.7
langchain-google-genai>=2.0.6
langchain-community>=0.3.7
faiss-cpu>=1.12.0
pydantic>=2.10.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
